backend/requirements.txt
bcrypt==3.2.0
cffi==1.15.0
click==8.0.3
cryptography==35.0.0
dnspython==2.1.0
email-validator==1.1.3
Flask==2.0.2
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
flask-mongoengine==1.0.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
greenlet==1.1.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.2
PyJWT==2.3.0
MarkupSafe==2.0.1
mongoengine==0.23.1
pycparser==2.20
PyJWT==2.3.0
pymongo==3.12.1
six==1.16.0
Werkzeug==2.0.2
WTForms==2.3.3
